fix: split GetLongRangeNodes into own file, fix query loop
